% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrGBM.R
\docType{data}
\name{LearnerRegrGBM}
\alias{LearnerRegrGBM}
\alias{mlr_learners_regr.gbm}
\title{Regression gbm Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:RegrClassif]{mlr3::RegrClassif}.}
\usage{
LearnerRegrGBM
}
\description{
A \link[mlr3:RegrClassif]{mlr3::RegrClassif} for a classification gbm implemented in \code{\link[gbm:gbm]{gbm::gbm()}} in package \CRANpkg{gbm}.
}
\keyword{datasets}
