% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learner.gbm-package}
\alias{mlr3learner.gbm}
\alias{mlr3learner.gbm-package}
\title{mlr3learner.gbm: gbm learner from gbm package for mlr3}
\description{
gbm learner from gbm package for mlr3.
}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

}
