% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learnergbm-package}
\alias{mlr3learnergbm}
\alias{mlr3learnergbm-package}
\title{mlr3learnergbm: gbm learner from gbm package for mlr3}
\description{
gbm learner from gbm package for mlr3.
}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

}
