% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifGBM.R
\docType{data}
\name{LearnerClassifGBM}
\alias{LearnerClassifGBM}
\alias{mlr_learners_classif.gbm}
\title{Classification gbm Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifGBM
}
\description{
A \link{LearnerClassif} for a classification gbm implemented in gbm::gbm()] in package \CRANpkg{gbm}.
}
\keyword{datasets}
